refactoring images to a new root level namespace
