+++ src/App/theme.ts
@@ -1,6 +1,6 @@
 import {Theme} from '@shopify/polaris/types/components/ThemeProvider';
 
-import {logo} from './images';
+import {logo} from 'images';
 
 const theme: Theme = {
   logo: {
